% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{blm}
\alias{blm}
\title{Fit a bayesian linear regression}
\usage{
blm(y, X)
}
\arguments{
\item{data, }{a list with the data in stan format}
}
\description{
Fit a bayesian linear regression
}
\examples{

n = 100
X = cbind(rnorm(n), rnorm(n))
beta = c(1, 2)
sigma = 0.01
y = X\%*\%beta + rnorm(n, 0, sigma)

}
