% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediascores-package.R
\docType{package}
\name{mediascores-package}
\alias{mediascores-package}
\title{mediascores: News-sharing Ideology from Social Media Link Data}
\description{
This library enables the examination of political communication by
politicians and ordinary users on social media by using news stories shared
by each (via URLs) to estimate the news-sharing ideology of those actors, and
the ideology of the content that they share itself.
}
\details{
The library provides three sets of functions for the following tasks:
(1) model estimation, (2) model diagnostics, and (3) model summary
}
\section{Model estimation}{


The core estimation function, \code{\link{mediascores}}, requires as input a
matrix or data.frame where the rows are social media users and the
columns are the URL domains shared by each user. Each cell of the matrix
represents the count of each domain shared by each user. If, for example, a
user shares 10 news stories from the New York Times, the value of the cell
indicated by the row of that user and the column of the New York Times would
be 10.
}

\section{Model diagnostics}{


The object outputted from the estimation function \code{\link{mediascores}}
is a \code{\link{stanfit}} object from the \href{https://mc-stan.org}{Stan Bayesian inference engine}. All methods
for analysis of the object useable for Stan objects can therefore used for
the objected outtputed from that function. A helper function
\code{\link{rhat}} is provided to give a quick summary of R-hat values to
assess convergence for the parameters of interest.
}

\section{Model summary}{


Point estimates and credible intervals for the quantities of interest
(i.e. the ideology of social media users and news organizations) are
accessible through the \code{\link{point_est}} function.
}

\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. \url{https://mc-stan.org}
}
