% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PolShares}
\alias{PolShares}
\title{News Media Domains Tweeted by members of the 116th US Congress and
other political actors (2016-2021)}
\format{
A data frame with 641 rows and 190 variables
}
\usage{
data(PolShares)
}
\description{
The counts of news media links tweeted or
retweeted by members of the 116th US Congress and other political actors.

\describe{
\item{name}{name of politician or political actors}
\item{role}{political role of politician (or organization) \{"House", "Senate", "Governor", "Politician-Other", "Delegate", "Other"\}}
\item{nominate_name}{name of members of Congress as specified by \url{voteview.com}}
\item{party}{political party of a politician (or organization) \{"Democrat", "Republican", "Independent"\}}
\item{affiliation}{political party a politician (or organization) is affiliated with \{"Democrat", "Republican"\}}
\item{group}{Democrat = 1, Republican = 2}
\item{nominate}{Nominate score of members of Congress (\url{voteview.com})}
\item{remaining variables}{remaining variables indicate the count for news shared to each news media domain}
}
}
