% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{rhat}
\alias{rhat}
\title{Display Rhat values for the parameters of interest}
\usage{
rhat(posterior, pars = NULL)
}
\arguments{
\item{posterior, }{object of class \code{rstan::\link[rstan]{stanfit}}.}

\item{pars}{character vector specifying parameters to extract point
estimates for. If \code{NULL} all parameters are selected.
Groups of parameters can be extracted by specifying the
parameter name without the index value. E.g. \code{"theta"} produces point
estimates for all \code{theta} parameter (i.e. \code{"theta[1]",
"theta[2]", ect.}). See \code{\link{mediascores}} for model details and run
   \code{names(posterior)} to see all available parameters in the model).}
}
\value{
Returns a vector of rhat values for the parameters specified by the pars
argument.
}
\description{
[TODO: EXTENDED HEADER]
}
\details{
[TODO: THIS IS THE DETAILS SECTION: FILL IN INFORMATION ON THE MODEL PARAMERS]
}
\examples{
\dontrun{
simulated_data <- simulate_data()
posterior <- mediascores(Y = simulated_data$Y, group = simulated_data$group,
                         anchors = simulated_data$anchors,
                         variational = TRUE)
rhat(posterior, pars = c("theta", "zeta"))
}
}
