% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{blm}
\alias{blm}
\title{Fit a bayesian linear regression}
\usage{
blm(y, X)
}
\arguments{
\item{y}{numeric, vector of outcomes}

\item{X}{matrix, matrix of predictors (1st col intercept)}
}
\description{
Fit a bayesian linear regression
}
\examples{

n = 100
X = cbind(1, rnorm(n), rnorm(n))
beta = c(0, 1, 2)
sigma = 0.01
y = as.numeric(X\%*\%beta + rnorm(n, 0, sigma))
blm(y, X) 
}
