% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediascores-package.R
\docType{package}
\name{mediascores-package}
\alias{mediascores-package}
\title{The 'mediascores' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. http://mc-stan.org
}
