% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediascores.R
\name{mediascores}
\alias{mediascores}
\title{Estimate the mediascores model}
\usage{
mediascores(Y, group = NULL, anchors, algorithm = c("sampling",
  "meanfield"), adapt_delta = 0.8, warmup = 1000, iter = 2000,
  chains = 4, refresh = 50)
}
\description{
Estimate the mediascores model
}
\examples{
sim_data <- simulate_data(200, 500)
posterior <- mediascores(sim_data$Y, sim_data$group, sim_data$anchors, 
                          algorithm = "meanfield")
}
