% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{point_est}
\alias{point_est}
\title{Provide a point estimate for each parameter of interest and its
credible interval}
\usage{
point_est(posterior, pars = c("theta", "theta_mu", "theta_sigma", "zeta",
  "alpha", "alpha_mu", "alpha_sigma", "gamma", "gamma_sigma",
  "omega_domain", "omega_user"), prob = 0.9)
}
\arguments{
\item{posterior, }{object of class \code{rstan::\link[rstan]{stanfit}}.}

\item{pars}{[TODO: FILL IN]. Document the parameter selection syntax. If 
too long for the argument documentation refer to details section and 
explain there}

\item{prob}{[TODO: FILL IN].}
}
\value{
Returns a three-columm matrix containing the median of the posterior of each
paramter of interest and a credible interval
}
\description{
[TODO: EXTENDED HEADER]
}
\details{
[TODO: THIS IS THE DETAILS SECTION: FILL IN INFORMATION ON THE MODEL PARAMERS]
}
\examples{
\dontrun{
simulated_data <- simulate_data()
posterior <- mediascores(Y = simulated_data$Y, group = simulated_data$group, 
                         anchors = simulated_data$anchors, 
                         variational = TRUE)
point_est(posterior, pars = c("theta", "theta_mu"), prob = 0.90)
}
}
