% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_estimation.R
\name{rhat}
\alias{rhat}
\title{Display Rhat values for the parameters of interest}
\usage{
rhat(posterior, pars = c("theta", "theta_mu", "theta_sigma", "zeta",
  "zeta_sigma", "alpha", "alpha_mu", "alpha_sigma", "gamma", "gamma_sigma",
  "omega_domain", "omega_user"))
}
\arguments{
\item{posterior, }{object of class \code{rstan::\link[rstan]{stanfit}}.}

\item{pars}{[TODO: FILL IN].}
}
\value{
Returns a vector of rhat values for the parameters specified by the pars
argument
}
\description{
[TODO: EXTENDED HEADER]
}
\details{
[TODO: THIS IS THE DETAILS SECTION: FILL IN INFORMATION ON THE MODEL PARAMERS]
}
\examples{
\dontrun{
simulated_data <- simulate_data()
posterior <- mediascores(Y = simulated_data$Y, group = simulated_data$group, 
                         anchors = simulated_data$anchors, 
                         variational = TRUE)
rhat(posterior, pars = c("theta", "zeta"))
}
}
